MINOR: channel: Report EOI on the input channel if it was reached in the mux
The flag CF_EOI is now set on the input channel when the flag CS_FL_EOI is set
on the corresponding conn_stream. In addition, if a read activity is reported
when this flag is set, the stream is woken up.

This patch should be backported to 1.9.

(cherry picked from commit 297d3e2e0f8f02636b9d58c40d7cab56a385f512)
Signed-off-by: Willy Tarreau <w@1wt.eu>