REGTEST: remove unexpected "nbthread" statement from Lua test cases
Lua test files 2 and 3 fail when threads are disabled because of a
"nbthread" statement that seems to be a leftover from an ancient
configuration. One of them even mentions a commit message showing
a reproducer not involving threads. Let's clean this up so that
running the tests without threads also works.

This should be backported to 1.9 as the problem also exists there.

(cherry picked from commit 26ecceaadc929368e1ee8086fa8ff9c71fd60c62)
Signed-off-by: Willy Tarreau <w@1wt.eu>