MINOR: tools: make memvprintf() never pass a NULL target to vsnprintf()
Most modern platforms don't touch the output buffer when the size
argument is null, but there exist a few old ones (like AIX 5 and
possibly Tru64) where the output will be dereferenced anyway, probably
to write the trailing null, crashing the process. memprintf() uses this
to measure the desired length.

There is a very simple workaround to this consisting in passing a pointer
to a character instead of a NULL pointer. It was confirmed to fix the issue
on AIX 5.1.

(cherry picked from commit e0609f5f49f55b122e7da9bd1d3b1b786366e80c)
[wt: it likely makes sense to backport this to all supported branches]
Signed-off-by: Willy Tarreau <w@1wt.eu>