BUG/MINOR: cli: correctly handle abns in 'show cli sockets'
The 'show cli sockets' was not handling the abns sockets. This is a
problem since it uses the AF_UNIX family, it displays nothing
in the path column because the path starts by \0.

Should be backported to 1.9 and 1.8.

(cherry picked from commit 75812a7a3cdc853ceee0b3fb2173db060057ba35)
Signed-off-by: Christopher Faulet <cfaulet@haproxy.com>