BUG/MEDIUM: spoe: Return an error if nothing is encoded for fragmented messages
If the maximum frame size is very small with a large message or argument name,
it is possible to be unable to encode anything. In such case, it is important to
stop processing returning an error otherwise we will retry in loop to encode the
message, failing each time because of the too small frame size.

This patch must be backported to 1.9 and 1.8.

(cherry picked from commit a715ea82eacf4ccf7f447bf4dd4111cc29fe171e)
Signed-off-by: William Lallemand <wlallemand@haproxy.org>