BUG/MEDIUM: h2: Revamp the way send subscriptions works.
Instead of abusing the SUB_CALL_UNSUBSCRIBE flag, revamp the H2 code a bit
so that it just checks if h2s->sending_list is empty to know if the tasklet
of the stream_interface has been waken up or not.
send_wait is now set to NULL in h2_snd_buf() (ideally we'd set it to NULL
as soon as we're waking the tasklet, but it can't be done, because we still
need it in case we have to remove the tasklet from the task list).

(cherry picked from commit 998410a41be37569a1e023c1402876fb3f3832ae)
[wt: while the backport didn't appear necessary at first, two recent
     reports involving the same parts tend to justify having it in 1.9]
Signed-off-by: Willy Tarreau <w@1wt.eu>