BUG/MINOR: mworker: don't exit with an ambiguous value
When the sigchld handler is called and waitpid() returns -1,
the behavior of waitpid() with the status variable is undefined.
It is not a good idea to exit with the value contained in it.

Since this exit path does not use the exitcode variable, it means that
this is an expected and successful exit.

This should be backported in 1.9, code has moved,
mworker_catch_sigchld() is in haproxy.c.

(cherry picked from commit 4cf4b33744d79fdc7c8d6a47b3c553c4367f36a1)
[wt: applied to haproxy.c]
Signed-off-by: Willy Tarreau <w@1wt.eu>