BUG/MINOR: http_fetch/htx: Allow permissive sample prefetch for the HTX
As for smp_prefetch_http(), there is now a way to successfully perform a
prefetch in HTX, even if the message forwarding already begun. It is used for
the sample fetches "req.proto_http" and "method".

This patch must be backported to 1.9.

(cherry picked from commit 5ec8bcb0218456f19d253f87ce5b4a483aa1e949)
Signed-off-by: Christopher Faulet <cfaulet@haproxy.com>